requirements.txt
mariadb
requests
pandas
pillow
discord.py
fake_useragent